Separate out __init__.py into better files
