+++ tracing/optimisation/enums.py
@@ -1,0 +1,7 @@
+import enum
+
+class TriggerStatus(enum.Enum):
+    INACTIVE = 0
+    ENTRY = 1
+    ONGOING = 2
+    EXIT = 3
